Start parameterizing grep commands
This changelist starts to parameterize each grep command via a
dictionary. This is a work in progress and some commands are not yet
parameterized 100%.
The longer term plan is to allow this dictionary to have custom entries
so that users can configure custom grep commands.
Add experimental support for 'git grep'
Add skeleton for g:EasyGrepFilesToInclude