Implement the g:EasyGrepRoot option
This option allows the user to control where the grep command is invoked from.
It defaults to the old behavior of invoking grep from the current directory. I
added support for the following options:
- cwd (the default)
- search (searches for a file or directory)
- arbitrary location

There is still more work to do to ensure that every part of the script respects
the root. In particular the GetFileTargetList needs to be updated and so these
options probably won't worry correctly until this is addressed.