Use ' for patternprefix and patternpostfix not "
Because bash interprets the command we pass it, escaping characters with
a single backslash when enclosed in double quotes doesn't change the
command the underlying grep program recieves.
If we enclose command arguments in single quotes, that disables bash
expansion for that argument.
This allows the escape backslash to be seen by the underlying grep
program.
I don't expect this will cause some escaping problems further down the
line, but with shellscript you can never tell.